agents/requirements.txt
# Google ADK
google-adk==1.0.0
google-genai[adk]==1.25.0

# Google Cloud
google-cloud-aiplatform==1.73.0

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.12.0

# Configuration
python-dotenv==1.1.0
pyyaml==6.0.1

# FastAPI (for agent server)
fastapi==0.129.0
uvicorn[standard]==0.41.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0